fixed measurement controller

+++ src/DotNetCoreWebApi/Controllers/MeasurementController.cs
@@ -29,6 +29,11 @@ namespace DotNetCoreWebApi.Controllers
         public async Task<IActionResult> Get(long id)
         {
             var measurement = await _measurementRepository.Get(id);
+            
+            if (measurement == null)
+            {
+                return NotFound("The Measurement record couldn't be found.");
+            }
 
             return Ok(measurement);
         }
@@ -36,6 +41,11 @@ namespace DotNetCoreWebApi.Controllers
         [HttpPost]
         public async Task<IActionResult> Post(Measurement measurement)
         {
+            if (measurement == null)
+            {
+                return BadRequest("Measurement is null.");
+            }
+
             await _measurementRepository.Add(measurement);
 
             return CreatedAtAction(nameof(Get), new { id = measurement.Id }, measurement);
@@ -45,6 +55,12 @@ namespace DotNetCoreWebApi.Controllers
         public async Task<IActionResult> Put(long id, Measurement measurement)
         {
             var measurementToUpdate = await _measurementRepository.Get(id);
+
+            if (measurementToUpdate == null)
+            {
+                return NotFound("The Measurement record couldn't be found.");
+            }
+
             await _measurementRepository.Update(measurementToUpdate, measurement);
 
             return NoContent();
@@ -54,6 +70,12 @@ namespace DotNetCoreWebApi.Controllers
         public async Task<IActionResult> Delete(long id)
         {
             var measurementToDelete = await _measurementRepository.Get(id);
+            
+            if (measurementToDelete == null)
+            {
+                return NotFound("The Measurement record couldn't be found.");
+            }
+
             await _measurementRepository.Delete(measurementToDelete);
 
             return NoContent();
